+++ doc/source/changelog.rst
@@ -3,6 +3,30 @@
 ChangeLog
 =========
 
+Version 4.2.0
+-------------
+
+This update to SOXS contains new features and a bugfix.
+
+* Installation and use on Windows 64-bit platforms is now supported.
+* New PSF models using encircled energy fraction (EEF) files are now supported.
+  See :ref:`psf-models` for more details.
+* The *XRISM* *Resolve* instrument specification has been updated, and a new
+  instrument specification for *Xtend* has been added. See :ref:`xrism` for
+  more details.
+* The *AXIS* instrument specification has been updated. See :ref:`axis` for
+  more details.
+* If one had not binned a :class:`~soxs.spectra.Spectrum` object more finely
+  than the instrument's ARF/RMF when using :func:`~soxs.instrument.simulate_spectrum`,
+  then gaps would appear in the resulting convolved spectrum. This is now
+  handled by linearly interpolating the spectral model into the ARF energy
+  bins.
+* The *LEM* instrumental background has been boosted to 1 counts/s/keV/(30'x30')
+  from the previous value of 0.07 counts/s/keV/(30'x30').
+* A new function for creating a SIMPUT catalog including models for the Galactic
+  foreground and the CXB point sources, :func:`~soxs.simput.make_bkgnd_simput`,
+  has been included. See :ref:`bkgnd-simput` for more details.
+
 Version 4.1.0
 -------------